+++ Client/src/Pages/Uptime/Monitors/index.jsx
@@ -4,7 +4,7 @@ import Greeting from "../../../Utils/greeting";
 import StatusBoxes from "./Components/StatusBoxes";
 import UptimeDataTable from "./Components/UptimeDataTable";
 import Pagination from "../../../Components/Table/TablePagination";
-import CreateMonitorHeader from "../../../Components/CreateMonitorHeader";
+import CreateMonitorHeader from "../../../Components/MonitorCreateHeader";
 import Fallback from "../../../Components/Fallback";
 import GenericFallback from "../../../Components/GenericFallback";
 import SearchComponent from "./Components/SearchComponent";
@@ -107,7 +107,7 @@ const UptimeMonitors = () => {
 		order: sort?.order,
 		triggerUpdate: monitorUpdateTrigger,
 	});
-	const totalMonitors = monitorsSummary?.totalMonitors ?? 0;
+	const totalMonitors = monitorsSummary?.totalMonitors;
 
 	if (networkError) {
 		return (
@@ -135,6 +135,7 @@ const UptimeMonitors = () => {
 			/>
 		);
 	}
+	console.log({ isLoading });
 
 	return (
 		<Stack
